+++ settings/Session.go
@@ -2,5 +2,5 @@ package settings
 
 // Branding contains the branding settings of the app.
 type Session struct {
-	Timeout uint `json:"timeout"`
+	Timeout uint64 `json:"timeout"`
 }
